chore: update `react-router-dom` package
