+++ src/utils/testing-library.tsx
@@ -2,11 +2,11 @@
 import { ARK } from "@ardenthq/sdk-ark";
 import { Contracts, Environment } from "@ardenthq/sdk-profiles";
 import { render, RenderResult } from "@testing-library/react";
-import { BrowserHistory, createHashHistory, HashHistory, To } from "history";
+import { createHashHistory, HashHistory, To } from "history";
 import React from "react";
 import { FormProvider, useForm, UseFormMethods } from "react-hook-form";
 import { I18nextProvider } from "react-i18next";
-import { Router, Routes } from "react-router-dom";
+import { Routes } from "react-router-dom";
 import { Context as ResponsiveContext } from "react-responsive";
 import { ConfigurationProvider, EnvironmentProvider, LedgerProvider, NavigationProvider } from "@/app/contexts";
 import { useProfileSynchronizer } from "@/app/hooks/use-profile-synchronizer";
@@ -18,6 +18,7 @@ import TestingPasswords from "@/tests/fixtures/env/testing-passwords.json";
 import DefaultManifest from "@/tests/fixtures/coins/ark/manifest/default.json";
 import { StubStorage } from "@/tests/mocks";
 import { connectedTransport as ledgerTransportFactory } from "@/app/contexts/Ledger/transport";
+import {CustomRouter} from "./CustomRouter";
 export {
 	mockNanoSTransport,
 	mockLedgerTransportError,
@@ -114,26 +115,6 @@ interface RenderWithRouterOptions {
 	profileSynchronizerOptions?: Record<string, any>;
 }
 
-interface Props {
-	basename?: string;
-	children: React.ReactNode;
-	history: BrowserHistory;
-}
-
-const CustomRouter = ({ basename, children, history }: Props) => {
-	const [state, setState] = React.useState({
-		action: history.action,
-		location: history.location,
-	});
-
-	React.useLayoutEffect(() => history.listen(setState), [history]);
-
-	return (
-		<Router basename={basename} location={state.location} navigator={history} navigationType={state.action}>
-			{children}
-		</Router>
-	);
-};
 
 const renderWithRouter = (
 	component: React.ReactElement,
